reactor: turn bench-poll into a proper benchmark
